requirements.txt
pandas
pytesseract
Pillow